refactor(quality): extract shared portfolio weights and finnhub earnings helpers
Removes two ~50-line duplicated blocks flagged by SonarCloud CPD:
- finnhub earnings calendar fetch (52 lines): MCP tooling now delegates to
  app/services/market_events/finnhub_helpers.fetch_earnings_calendar_finnhub
- portfolio weight computation (55 lines): _round_pct/_build_weights extracted
  from PortfolioDecisionService and PortfolioPositionDetailService into
  app/services/portfolio_weights as pure module functions

No behavior change. All affected tests updated to call the module function
directly; relevant test suites pass (285 tests).
