+++ app/mcp_server/tooling/fundamentals_sources_finnhub.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import asyncio
-import datetime
 import logging
 from typing import Any
 
@@ -177,58 +176,11 @@ async def _fetch_earnings_calendar_finnhub(
     from_date: str | None = None,
     to_date: str | None = None,
 ) -> dict[str, Any]:
-    client = _get_finnhub_client()
+    from app.services.market_events.finnhub_helpers import (
+        fetch_earnings_calendar_finnhub,
+    )
 
-    if not from_date:
-        from_date = datetime.date.today().isoformat()
-    if not to_date:
-        to_date = (datetime.date.today() + datetime.timedelta(days=30)).isoformat()
-
-    def fetch_sync() -> dict[str, Any]:
-        return client.earnings_calendar(
-            symbol=symbol.upper() if symbol else "",
-            _from=from_date,
-            to=to_date,
-        )
-
-    result = await asyncio.to_thread(fetch_sync)
-
-    if not result or not result.get("earningsCalendar"):
-        return {
-            "symbol": symbol,
-            "instrument_type": "equity_us",
-            "source": "finnhub",
-            "from_date": from_date,
-            "to_date": to_date,
-            "count": 0,
-            "earnings": [],
-        }
-
-    earnings = []
-    for item in result.get("earningsCalendar", []):
-        earnings.append(
-            {
-                "symbol": item.get("symbol", ""),
-                "date": item.get("date"),
-                "hour": item.get("hour", ""),
-                "eps_estimate": item.get("epsEstimate"),
-                "eps_actual": item.get("epsActual"),
-                "revenue_estimate": item.get("revenueEstimate"),
-                "revenue_actual": item.get("revenueActual"),
-                "quarter": item.get("quarter"),
-                "year": item.get("year"),
-            }
-        )
-
-    return {
-        "symbol": symbol,
-        "instrument_type": "equity_us",
-        "source": "finnhub",
-        "from_date": from_date,
-        "to_date": to_date,
-        "count": len(earnings),
-        "earnings": earnings,
-    }
+    return await fetch_earnings_calendar_finnhub(symbol, from_date, to_date)
 
 
 __all__ = [
