+++ app/services/portfolio_decision_service.py
@@ -13,6 +13,7 @@ from app.mcp_server.tooling.fundamentals._support_resistance import (
 from app.mcp_server.tooling.market_data_quotes import _get_indicators_impl
 from app.models.portfolio_decision_run import PortfolioDecisionRun
 from app.schemas.portfolio_decision import PortfolioDecisionSlateResponse
+from app.services.portfolio_weights import build_weights as _build_portfolio_weights
 
 logger = logging.getLogger(__name__)
 
@@ -174,7 +175,7 @@ class PortfolioDecisionService:
             journal = journals.get(symbol)
             context = contexts.get(position_key) or self._empty_market_context()
 
-            weights = self._build_weights(positions, p)
+            weights = _build_portfolio_weights(positions, p)
             group = self._build_symbol_group(p, journal, context, weights)
             symbol_groups.append(group)
             self._add_group_to_summary(summary_counts, market_type, group)
@@ -792,59 +793,3 @@ class PortfolioDecisionService:
 
     def _default_execution_boundary(self) -> dict[str, Any]:
         return dict(DEFAULT_EXECUTION_BOUNDARY)
-
-    def _round_pct(self, value: float | None) -> float | None:
-        if value is None:
-            return None
-        return round(value, 1)
-
-    def _build_weights(
-        self,
-        positions: list[dict[str, Any]],
-        base: dict[str, Any],
-    ) -> dict[str, float | None]:
-        def get_eval_krw(p: dict) -> float | None:
-            market_type = str(p.get("market_type") or "").upper()
-            val = p.get("evaluation_krw")
-            if val is not None:
-                return float(val)
-            if market_type == "US":
-                return None
-            return float(p.get("evaluation", 0) or 0)
-
-        base_evaluation_krw = get_eval_krw(base)
-        if base_evaluation_krw in (None, 0):
-            return {"portfolio_weight_pct": None, "market_weight_pct": None}
-
-        portfolio_values = [get_eval_krw(p) for p in positions]
-        total_portfolio_eval_krw = (
-            sum(value for value in portfolio_values if value is not None)
-            if all(value is not None for value in portfolio_values)
-            else None
-        )
-
-        market_type = base.get("market_type")
-        same_market_values = [
-            get_eval_krw(p) for p in positions if p.get("market_type") == market_type
-        ]
-        total_same_market_eval_krw = (
-            sum(value for value in same_market_values if value is not None)
-            if all(value is not None for value in same_market_values)
-            else None
-        )
-
-        portfolio_weight = (
-            (base_evaluation_krw / total_portfolio_eval_krw) * 100
-            if total_portfolio_eval_krw not in (None, 0)
-            else None
-        )
-        market_weight = (
-            (base_evaluation_krw / total_same_market_eval_krw) * 100
-            if total_same_market_eval_krw not in (None, 0)
-            else None
-        )
-
-        return {
-            "portfolio_weight_pct": self._round_pct(portfolio_weight),
-            "market_weight_pct": self._round_pct(market_weight),
-        }
