+++ app/services/portfolio_position_detail_service.py
@@ -10,6 +10,7 @@ from app.mcp_server.tooling.fundamentals._valuation import (
 )
 from app.mcp_server.tooling.market_data_quotes import _get_indicators_impl
 from app.mcp_server.tooling.orders_history import get_order_history_impl
+from app.services.portfolio_weights import build_weights as _build_portfolio_weights
 
 
 class PortfolioPositionDetailNotFoundError(Exception):
@@ -51,7 +52,7 @@ class PortfolioPositionDetailService:
             current_price=base.get("current_price"),
         )
 
-        weights = self._build_weights(positions, base)
+        weights = _build_portfolio_weights(positions, base)
         indicators = await self._fetch_action_inputs(
             market_type=str(market_type).upper(), symbol=symbol
         )
@@ -84,62 +85,6 @@ class PortfolioPositionDetailService:
             "journal": journal,
             "weights": weights,
             "action_summary": action_summary,
-        }
-
-    def _round_pct(self, value: float | None) -> float | None:
-        if value is None:
-            return None
-        return round(value, 1)
-
-    def _build_weights(
-        self,
-        positions: list[dict[str, Any]],
-        base: dict[str, Any],
-    ) -> dict[str, float | None]:
-        def get_eval_krw(p: dict) -> float | None:
-            market_type = str(p.get("market_type") or "").upper()
-            val = p.get("evaluation_krw")
-            if val is not None:
-                return float(val)
-            if market_type == "US":
-                return None
-            return float(p.get("evaluation", 0) or 0)
-
-        base_evaluation_krw = get_eval_krw(base)
-        if base_evaluation_krw in (None, 0):
-            return {"portfolio_weight_pct": None, "market_weight_pct": None}
-
-        portfolio_values = [get_eval_krw(p) for p in positions]
-        total_portfolio_eval_krw = (
-            sum(value for value in portfolio_values if value is not None)
-            if all(value is not None for value in portfolio_values)
-            else None
-        )
-
-        market_type = base.get("market_type")
-        same_market_values = [
-            get_eval_krw(p) for p in positions if p.get("market_type") == market_type
-        ]
-        total_same_market_eval_krw = (
-            sum(value for value in same_market_values if value is not None)
-            if all(value is not None for value in same_market_values)
-            else None
-        )
-
-        portfolio_weight = (
-            (base_evaluation_krw / total_portfolio_eval_krw) * 100
-            if total_portfolio_eval_krw not in (None, 0)
-            else None
-        )
-        market_weight = (
-            (base_evaluation_krw / total_same_market_eval_krw) * 100
-            if total_same_market_eval_krw not in (None, 0)
-            else None
-        )
-
-        return {
-            "portfolio_weight_pct": self._round_pct(portfolio_weight),
-            "market_weight_pct": self._round_pct(market_weight),
         }
 
     async def _fetch_action_inputs(
